int32 in v1

+++ vllm/v1/kv_offload/base.py
@@ -400,7 +400,7 @@ class BlockIDsLoadStoreSpec(LoadStoreSpec, ABC):
     """
 
     def __init__(self, block_ids: list[int]):
-        self.block_ids = np.array(block_ids, dtype=np.int64)
+        self.block_ids = np.array(block_ids, dtype=np.int32)
 
     def __repr__(self) -> str:
         return repr(self.block_ids)
